backend-vercel/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.31
prometheus-client==0.20.0
requests==2.31.0
scikit-learn==1.5.0
python-dotenv==1.0.1
flask-cors==4.0.1
pytest==7.4.4
gunicorn==22.0.0